% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclipseplot.R
\name{get_eclipse_data}
\alias{get_eclipse_data}
\title{Get path to example files (.csv)}
\usage{
get_eclipse_data(name = "brief")
}
\arguments{
\item{name}{Name of the file ("brief" or "long").}
}
\value{
A character string containing the absolute file path to the requested
sample CSV file within the package installation directory.
}
\description{
Get path to example files (.csv)
}
