% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messy}
\alias{messy}
\title{Mock messy data for eclipsedf testing}
\format{
A data frame with 4 rows and 13 variables of varied naming conventions.
}
\usage{
messy
}
\description{
A dataset containing messy headers and non-standard values to test the
cleaning and standardization capabilities of the eclipsedf function.
}
\keyword{datasets}
