% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_brief}
\alias{sample_brief}
\title{Sample brief dataset for eclipseplot}
\format{
A data frame with 5 rows and 13 variables:
\describe{
  \item{pmid}{PubMed ID}
  \item{study}{Study citation label}
  \item{item1_step1, item1_step2, item2_step1, item2_step2, item3_step1, item3_step2, item4_step1, item4_step2, item5_step1, item5_step2, item6_step2}{Risk of bias for the core items}
}
}
\usage{
sample_brief
}
\description{
A small dataset containing risk of bias assessments for testing.
}
\keyword{datasets}
