% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_long}
\alias{sample_long}
\title{Sample long dataset for eclipseplot}
\format{
A data frame with 26 rows and 13 variables.
}
\usage{
sample_long
}
\description{
A larger dataset with more studies and optional items.
}
\keyword{datasets}
